requirements.txt
# Core dependencies
anthropic>=0.39.0
requests>=2.31.0
beautifulsoup4>=4.12.0
faker>=20.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Database
sqlite3  # Built-in with Python

# Scraping
lxml>=4.9.0
html5lib>=1.1

# Date handling
python-dateutil>=2.8.2

# Logging
colorlog>=6.7.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0